refactor: `react-no-useless-fragment` always allow expressions
