+++ packages/eslint-plugin-jsx/src/rules/no-useless-fragment.md
@@ -72,22 +72,3 @@ const cat = <>meow</>
 
 {showFullName ? fullName : firstName}
 ```
-
-## Rule Options
-
-### `allowExpressions`
-
-When `true` single expressions in a fragment will be allowed. This is useful in
-places like Typescript where `string` does not satisfy the expected return type
-of `JSX.Element`. A common workaround is to wrap the variable holding a string
-in a fragment and expression.
-
-Examples of **correct** code for the rule, when `"allowExpressions"` is `true`:
-
-```jsx
-<>{foo}</>
-
-<>
-  {foo}
-</>
-```
