+++ packages/jsx/src/fragment.ts
@@ -1,8 +1,6 @@
 import { is, isOneOf, NodeType } from "@eslint-react/ast";
 import { type TSESTree } from "@typescript-eslint/types";
 
-import { isLiteral, isPaddingSpaces } from "./textnode";
-
 export const isFragment = (
   node: TSESTree.Node,
   pragma: string,
@@ -39,47 +37,4 @@ export function isFragmentElement(node: TSESTree.JSXElement, pragma: string, fra
     && name.object.type === NodeType.JSXIdentifier
     && name.object.name === pragma
     && name.property.name === fragment;
-}
-
-/**
- * Check if a JSXElement or JSXFragment has only one literal child and is not a child
- * @param node The AST node to check
- * @returns `true` if the node has only one literal child and is not a child
- * @example Somehow fragment like this is useful: <Foo content={<>ee eeee eeee ...</>} />
- */
-export function isFragmentWithOnlyTextAndIsNotChild(node: TSESTree.JSXElement | TSESTree.JSXFragment) {
-  return node.children.length === 1
-    && isLiteral(node.children[0])
-    && !(node.parent.type === NodeType.JSXElement || node.parent.type === NodeType.JSXFragment);
-}
-
-function containsCallExpression(node: TSESTree.Node) {
-  return node.type === NodeType.JSXExpressionContainer
-    && node.expression.type === NodeType.CallExpression;
-}
-
-/**
- * Check if a JSXElement or JSXFragment has less than two non-padding children and the first child is not a call expression
- * @param node The AST node to check
- * @returns boolean
- */
-export function isFragmentHasLessThanTwoChildren(node: TSESTree.JSXElement | TSESTree.JSXFragment) {
-  const nonPaddingChildren = node.children.filter(
-    (child) => !isPaddingSpaces(child),
-  );
-
-  if (nonPaddingChildren.length === 1) {
-    return !containsCallExpression(nonPaddingChildren[0] as TSESTree.Node);
-  }
-
-  return nonPaddingChildren.length === 0;
-}
-
-export function isFragmentWithSingleExpression(node: TSESTree.JSXElement | TSESTree.JSXFragment) {
-  const children = node.children.filter((child) => !isPaddingSpaces(child));
-
-  return (
-    children.length === 1
-    && children[0]?.type === NodeType.JSXExpressionContainer
-  );
 }
